kernel: Add 400 Hz timer frequency option
300 Hz is not an ideal timer frequency. While it is a decent balance in
terms of frequency, the length of a jiffy is irrational and thus rounded
in the kernel. Thus, one jiffy becomes 3.333333 ms, which can cause
rounding errors and imprecise scheduling, especially when desired
durations are rounded to 5 as are most hardcoded durations used in the
kernel.

400 Hz is just marginally higher while offering a rational jiffy length
of 2.5 ms rather than an odd and irrational length. This should fix many
theoretical issues in drivers and code across the kernel, such as one
that occurred in WALT (the load-tracking mechanism for EAS as of Android
4.4 and 4.9) which was fixed by CAF in commit 578db5dbecfa.

Furthermore, 400 Hz is a decent balance between 300 Hz (the current
"ideal" value for mobile devices) and 500 Hz (supposedly better, but may
cause slightly more battery drain and thermal output) that should be
able to marginally improve performance for some applications.

400 Hz is also cited as a tuning value by custom kernel developers in
search for the "perfect" frequency on the XanMod forums:
https://forum.xanmod.org/archive/index.php/thread-28.html

Measured values from a running system:
Before:
  # awk '{print$22/'$(tail -n 1 /proc/uptime|cut -d. -f1)"}" /proc/self/stat
  300.033

After:
  # awk '{print$22/'$(tail -n 1 /proc/uptime|cut -d. -f1)"}" /proc/self/stat
  400.097

Test: device boots, everything works, exposed frequency is different

Change-Id: I0e182d90f5551f9adf114550d67c137d5e9eab6e
Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: Tiktodz <dotkit@electrowizard.me>